Added 'self' to prop get/set + bug fixes
Add 'self' as the first argument of the generated property getter and
setter functions.

Fixed the code that generates the upcast macro so that vtable_offset
table is actually used. Also moved the vtable offset table to the beginning
of the class object, so one less ADD operation is needed for every upcast
operation.

Fixed an issue so that comments can now be added between function arguments.