Major restructure of zco compiler
Create a dispatch table per class. This table contains the
virtual function pointers for the class and all of its parent
classes. This table is kept separate from the global table
(which contains the static data members of the class) so that
we can share the same static data members across other objects
that are of a derived type.

Added support for protected data and methods.

Replace the use of string_t with ZString.