Initial commit for signal support
Added the ZClosureMarshal class. Added a signal map inside ZObject,
which will be used to do a lookup a vector of closures by a signal
name. Also changed the zco compiler so that a function name ending
in '!' will indicate that it is a signal and does not require a
body.