Add support for virtual data members and more...
Virtual data members are similar to global data members except that
their values are unique for an object of the class type and for an
object of a derived class type. Global data members are different
such that their values are the same for the class itself and for all
derived types.

This commit also includes more changes to ZCClosureMarshal, which
adds a dependency to libffi to handle calling conventions. A test
driver for signal is added. Also added a method map to ZObject so
methods can be looked up by their name (as string).

Also added a new storage mode to ZVector so that callbacks can be
called in a similar way to how ZMap calls them.