Added the event dispatcher
The ZEventLoop is a very simple class that acts as a dispatcher for
framework events. Currently, there are two events in the class:
draw_event, io_event. The application layer will raise the event
when signaled from the client. Interested objects (ie. the toplevel
window) can connect to these events to get notifications.

In this commit, the top level window gets notification of the draw_event
and it then sends back an example message.

This commit also includes a fix in the signal implementation with ZObjects;
The first argument from the caller was not dropped so the receiver of the
event got both the target ZObject and the source ZObject as the first two
arguments. It should've gotten just the target ZObject.

The test driver for signals is improved to account for this issue.