Reduce number of mallocs
Prior to this change, the get_type() of a class performed several
memory allocations based on the number of classes it is deriving from
(or the number of interfaces it is implementing). The number of
allocations have been reduced so that the final size is first calculated
and then a single malloc() is called.