Optimize ZVector capacity growth rate
Prior to this change, the capacity of a ZVector grew only to the
size requirement. This caused a lot of allocations to be performed
when pushing many items to a vector one at a time. Settng the
capacity to double the size requirement will greatly reduce the
number of the times the capacity has to be changed, thus reducing
the number of allocations needed.

This performance issue was realized after enabling the default
object tracker aka garbage collector. With this change, the
performance is still about 60% slower than the original (without GC).