Started working on to recycling old objects.
Started working on to recycling old objects stored in the object
tracker pool. The zco compiler has been modified so that the
generated class constructor first checks the object tracker if
it can provide an object. If not, it calls the 'new' function
to create and initialize a new object.