Support for vectors with multiple segments
This implementation works with test drivers as long as
the minimum segment size is 26 bytes. Having a smaller
sized segment exposes memory corruptions and leaks. These
leaks will be addressed in later iterations.

In this change, a new assign() method is added to ZVectorIter
so that one iterator's state can easily be copied to another.
ZVectorIter is also modified to be aware of mutiple segments.