Only use public APIs for vector segments
Prior to this change, ZVectorIter used public data members of vector segments
to get the 'prev' and 'next' pointer. In order to put all the reference counting
logic for prev/next in one place, the public properties are created for use by
other objects (such as ZVectorIter). These public properties handle the reference
counting.

This commit also contains the change in ZVectorIter to use the public APIs in
ZVectorSegment.