Use a static vtable to track offsets of Classes
Instead of using a struct member to get a reference to a parent class struct,
we now use a static vtable that maps type id to an offset. We can use this
offset to compute the address of any parent class within a Class struct, given
the parent class type id. This has the advantage of being able to get the
parent of the parent.