Create three types of allocators
Each one of these allocators are (will be) designed for a specific purpose.
The 'sys' allocator is just a pass through - it uses the malloc/free that
is provided by underlying C libray. The 'slab' allocator will use bins to
hold a group of blocks of the same size. This allocator is designed to
provide fixed-sized memory blocks very fast. The 'flex' allocator will
allocate memory in regions such that it can easily increase in size.

The stubs for these allocators, which merely calls malloc/free, are registered
with the context.

The vector segment class is changed so that if it is assigned an allocator,
it will use it. Otherwise, it will fall back to malloc/free. The vector class
will pass its own allocator to its segment if it's able to. If not, it will
get the 'flex' allocator from the context and pass it to the segment constructor.