Several bugfixes
Do not create internal vector when resetting a map or a string
Do not initialize allocator in init(). It will be initialized in the
    new() function of the class.
Notify object tracker of object destruction only if there is a reference to a
    tracker.
Added helper macros ALLOCATOR_FROM_OBJECT_OR_* to help get the allocator of
    an object or return a default fallback.
Release memory allocators once all other ZObjects are released.