Use a two-phase destruction process
Set 'flex' as the default allocator for vector segments, vectors,
maps, and strings.

Use a two-phase destruction process so that the objects that require
an allocator are destroyed first, and then the objects that don't require
it are destroyed last. This allows us to destroy the objects without
the objects knowing if it was allocated using a custom allocator or now.