Resize segments when possible
Vector segments can be resized if the allocator reports that the block
of memory can be resized in place.

This commit also changes the pointers in the context that point to allocators:
ctx->slab_allocator is now ctx->fixed_allocator
ctx->flex_allocator is still ctx->flex_allocator

This will prevent confusion between different terms and allocator names.