Multithread support
This commit adds the three new classes:
ZEventLoop - Manages the event loop associated with a worker thread
ZBind - Closure with argments. This object is passed cross-thread
    to schedule a task from one thread to another.
ZThreadSimpleMemoryAllocator - This is the thread-safe version of
    ZSimpleMemoryAllocator. This allocator is used when allocating
    memory during the post_task operation.

The API for post_task takes in a timeout period for which the task
should execute. The current implementation ignores the period and
assumes that it is 0. This will be fixed at a later commit.