Execute tasks according to its scheduled time
The current implementation uses the realtime clock to normalize the
schedule time for the tasks and to check when a task should be executed.
This method is suceptable to clock skews if someone changes he system
time while the process is running. This clock skew can lead to two
outcomes:
 - A burst of tasks to run that are scheduled to run later
 - Some tasks many run later than they are scheduled to run

A possible solution is to use a coarse monotonic clock to detect
clock skew. If a clock skew is detected, the schedule time for every
task will be adjusted based on the detected change in the system
time. This logic is not yet implemented.