Add z_bind_assign and z_string_iter_assign
This will allow users of ZBind to reuse a bind object by making a
copy of it, modifying it, disposing it, and then switching to use
the copy the next time around.

The z_string_iter_assign's case is very standard; Sometime it's
important to be able to copy iterators.