ZBind: Variable size object support
A few methods are added to ZBind to allow the user to add arguments to the
variable argument list of the target function.

The append_variable_buffer and unpack_variable_buffer methods allow the user to
push/pop a struct from the stack. It works by splitting the variable buffer
into chunks of 32-bit or 64-bit integers when pushing it onto the stack and
then rejoining them when poping it from the stack.