Fix deadlock
In the epoll implementation of ZEventLoop, the timeout was improperly
calculated so it would wait an unreasonable amount of time before
continuing to pump messages. This would cause a temporary deadlock
that would last a very long time.

This commit also refactors much of the code so that the conditional
variable implementation and the epoll implementation is cleanly
separated from one another.