Major refactor
This change-set reduces the dependency of many classes from the bootstrap
library by wrapping unused features behind preprocessor conditions.

This change also moves some classes from the core library to extra library.
The classes that are moved are not used by the zco meta-compiler.

Many fields from the zco_context_t structure are removed. These fields are
related to the event loop, the default marshaller and the event framework
manager. The event loop is moved to the extra library and is made into a
singleton under the context. The default marshaller no longer exist and
any object that is expected to be called by libffi must have a reference
to a closure marshaller.

The application will now create and destroy a zco_app_context instead of a
zco_context. This will allow some level of abstraction from the core
zco_context, such as the automatic destruction of the event loop if the
application forgets to destroy it.