Reimplement ZServer
Instead of using a nested event loop for watching the server socket and client
sockets, ZServer asks the event loop to notify it when an activity occurs on
a socket. The requires the event loop to support IO multiplexing but allows
all multiplexing logic to be in the event loop and greatly simplifies the
implementation of the ZServer class.

This change also allows the handlers of socket activity for client peers and
server peers to be very similar, implementation-wise. The ZServer acts as a
broker to instantiate a server peer to a TCP connection.

Added ZInOut which acts as a wrapper for file descriptors. This allows the file
descriptors to be ref-counted. It's possible to dup() the file descriptor to
add a reference to it but that requires a call to the kernel, something we
want to avoid.

Lazily remove fd subscriptions to avoid race conditions. This ensures that the
subscription list is not altered while it is being iterated.

Add commonly used callbacks to ZSocket. The application can install callbacks
to get notified when the particular event occurs on the socket. The callbacks
can be installed on a client peer or server peer of a connection.

Remove the support for dispatching socket callbacks to other event loops. The
ZSocket class now accepts callbacks to be installed so it cannot be serialized.
The serialization of the socket is a requirement for dispatching it across
event loops. The cross-eventloop dispatch wasn't even a used feature so it's
safe to remove it.

Added a simple test driver for network operations

Fixed memory leak in ZMap

Renamed ZIOMultiplexer to ZInOutMultiplexer