Use macros
Instead of generating all C code using the ZCO preprocessor, we now define
macros in zco-context.h for all common operations needed for the ZCO classes.
The ZCO preprocessor is modified to generate code to call the macros instead
of generating all the code that are in the macros. This will make it easier
for a developer to write a class without completely depending on the ZCO
preprocessor. Also, this will reduce the complexity in the ZCO preprocessor.