Pass the allocator directly to the callback
Prior to this commit, the subscription list would be provided with an
uncounted reference to the ZEventLoopLinux object and it would get the
allocator out of it. Even though the subscription list is managed by the
ZEventLoopLinux object, it's possible for the garbage collector to pick up the
ZEventLoopLinux object before the subscription list. Given that the
destruction of the subscription list depends on the ZEventLoopLinux object, it
would fail.

Since the memory allocator is relatively long lived compared to other objects,
it's safer to pull out the memory allocator and pass an uncounted reference to
the subscription list destruction callback.