Support insert into an empty vector
If insert() is called with an iterator that is pointing to a NULL segment and
the target vector is empty, it's likely that the get_begin() was called on the
vector when the vector was still empty. To handle this case, we create a
segment and then perform the insertion operation.