Added toolkit library
This library provides an abstration layer on top of commonly used classes. If
the runtime context is configured to run in NATIVE mode, the layer passes the
calls through. If the runtime context is configured to run in JS mode (as an
example scripting language), the layer generates the corresponding JS code and
stores it as a string within the runtime context. This JS code can later be
sent to a JS engine to run.