Start the renaming process from ZCO to Honeycomb
Originally, the project was named ZCO (Zero Copy Object) and the library was
named ZCLib (Zero Copy Library) because the primary decision making factor
every step of the way was to reduce the number of copies of data. Any data
that needs to be copied is lazily copied with copy-on-write.

The name ZCO is not a very attractive name and it describes an attribute of the
project that is an implementation detail. It's actually not an accurate
description because objects are not really zero-copied; They are virtually
copied and the copy operation is lazily performed when a write operation
happens. A more appropiate name for the project should emphasize the space
efficiency of the data structures but should not state an attribute that is
not entirely true.

The name Honeycomb describes certain features of the library. For example, the
hexagon shape of a honeycomb allows the maximum amount of honey storage per
cell. Since an initial design goal of the project is to reduce the number of
copies of data as much as possible, it is geared for space efficiency. It makes
sense to compare the space efficiency of the data structure with honeycomb.

All cells in a honeycomb are equal-sized, or at least as close to equal-sized
as possible. This is because honey bees parallelize their work very
efficiently with minimal (if any) contention between them. In this library,
almost all operations are thread-unsafe and they operate under a context that
is localized in a specific thread. This allows all tasks scheduled in a thread
to run as fast as possible without any wait on another thread. The event loop
is also optimized to stay in user space as much as possible when reloading
the work queue and only jumping into the kernel when absolutely necessary.